hotfix elementId

+++ BatchExport/Views/Params/ParametersTable.cs
@@ -5,7 +5,7 @@ namespace AlterTools.BatchExport.Views.Params
     public class ParametersTable
     {
         public string ModelName { get; set; }
-        public int ElementId { get; set; }
+        public long ElementId { get; set; }
         public Dictionary<string, string> Parameters { get; set; }
     }
 }
